# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Garimpei API
  description: |
    API do Garimpei — curadoria inteligente de produtos de afiliados Shopee.
    Backend Go servindo JSON. Autenticação via Firebase Auth (Bearer token).
  version: 1.0.0
  contact:
    name: Fernando Marques
servers:
  - url: https://garimpo-api-vj6afttbza-rj.a.run.app
    description: Produção (Cloud Run)
  - url: http://localhost:8080
    description: Desenvolvimento local
security:
  - BearerAuth: []
...
paths:
...
  /api/candidatos:
    get:
      tags: [Curadoria]
      summary: Lista candidatos ranqueados por teor
      security: []
      parameters:
        - name: estrategia
          in: query
          schema: { type: string, enum: [nicho, diversificada], default: nicho }
        - name: top
          in: query
          schema: { type: integer, default: 10 }
        - name: keyword
          in: query
          schema: { type: string }
        - name: categoria
          in: query
          schema: { type: string }
        - name: cat
          in: query
          description: ID de categoria Shopee
          schema: { type: integer }
        - name: comissao_min
          in: query
          schema: { type: number, example: 0.07 }
        - name: vendas_min
          in: query
          schema: { type: integer, example: 5 }
        - name: nota_min
          in: query
          schema: { type: number, example: 4.0 }
        - name: fonte
          in: query
          schema: { type: string, enum: [shopee, shopee-shop, csv] }
        - name: shop_ids
          in: query
          description: IDs de lojas separados por vírgula (fonte=shopee-shop)
          schema: { type: string, example: "123456,789012" }
        - name: sem_filtro
          in: query
          description: Se true, desliga filtros de elegibilidade
          schema: { type: boolean }
        - name: exploracao
          in: query
          description: Fração de exploração (0-0.9)
          schema: { type: number }
      responses:
        '200':
          description: Lista ranqueada
          content:
            application/json:
              schema:
                type: object
                properties:
                  estrategia: { type: string }
                  candidatos:
                    type: array
                    items: { $ref: '#/components/schemas/Candidato' }
                  total_bruto: { type: integer }
  /api/comparar:
    get:
      tags: [Curadoria]
      summary: Compara rankings nicho vs diversificada
      security: []
      parameters:
        - name: top
          in: query
          schema: { type: integer, default: 8 }
        - name: keyword
          in: query
          schema: { type: string }
      responses:
        '200':
          description: Ambos os rankings
          content:
            application/json:
              schema:
                type: object
                properties:
                  nicho: { type: array, items: { $ref: '#/components/schemas/Candidato' } }
                  diversificada: { type: array, items: { $ref: '#/components/schemas/Candidato' } }
  /api/eventos:
    post:
      tags: [Curadoria]
      summary: Registra evento de curadoria (seleção)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Evento'
      responses:
        '202':
          description: Evento registrado
  /api/buscas:
    get:
      tags: [Curadoria]
      summary: Lista buscas salvas do usuário
      responses:
        '200':
          description: Lista de buscas
          content:
            application/json:
              schema:
                type: object
                properties:
                  buscas:
                    type: array
                    items: { $ref: '#/components/schemas/Busca' }
                  store: { type: string }
    post:
      tags: [Curadoria]
      summary: Salvar/atualizar busca
      parameters:
        - name: remover
          in: query
          description: Se presente, marca busca como inativa (tombstone)
          schema: { type: boolean }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Busca' }
      responses:
        '202':
          description: Busca salva
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string }
                  id: { type: string }
                  ativo: { type: boolean }
  /api/lojas:
    get:
      tags: [Lojas]
      summary: Lista lojas monitoradas do usuário
      responses:
        '200':
          description: Lista de lojas
          content:
            application/json:
              schema:
                type: object
                properties:
                  lojas:
                    type: array
                    items: { $ref: '#/components/schemas/Busca' }
    post:
      tags: [Lojas]
      summary: Adicionar loja ao monitoramento
      description: |
        Aceita URL da Shopee (shopee.com.br/shop/{id}) ou ID numérico (5-15 dígitos).
        Cria uma busca com shop_ids e registra job no scheduler (cron padrão: a cada 4h).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [input]
              properties:
                input:
                  type: string
                  description: URL da loja ou ID numérico
                  examples:
                    - "https://shopee.com.br/shop/123456"
                    - "123456"
                cron:
                  type: string
                  description: Expressão cron (padrão "0 */4 * * *")
                  example: "0 */4 * * *"
                origem_padrao:
                  type: string
                  description: País de origem dos produtos da loja (ex. Coreia, Japão). Herdado por todos os produtos.
                  example: "Coreia"
      responses:
        '201':
          description: Loja adicionada
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, example: "ok" }
                  id: { type: string, example: "loja-123456" }
                  shop_id: { type: integer, example: 123456 }
                  nome: { type: string, example: "SKIN1004 Official" }
        '400':
          description: Input inválido
        '409':
          description: Loja já está sendo monitorada
    delete:
      tags: [Lojas]
      summary: Remover monitoramento de loja
      parameters:
        - name: id
          in: query
          required: true
          schema: { type: string }
          example: "loja-123456"
      responses:
        '200':
          description: Loja removida (tombstone)
        '404':
          description: Loja não encontrada
  /api/lojas/novidades:
    get:
      tags: [Lojas]
      summary: Novidades e variações de preço de uma loja
      parameters:
        - name: busca_id
          in: query
          schema: { type: string }
        - name: dias
          in: query
          schema: { type: integer, default: 7 }
      responses:
        '200':
          description: Novidades da loja
          content:
            application/json:
              schema: { $ref: '#/components/schemas/NovidadesLojas' }
  /api/lojas/evolucao:
    get:
      tags: [Análise]
      summary: Evolução de preço das lojas monitoradas
      description: Série temporal de preço médio por dia, com resumo global e top variações.
      parameters:
        - name: dias
          in: query
          schema: { type: integer, default: 30 }
      responses:
        '200':
          description: Dados de evolução
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EvolucaoLojas' }
...
  /api/alertas/testar:
    post:
      tags: [Alertas]
      summary: Enviar alerta de teste ou verificar alertas de uma loja
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                busca_id:
                  type: string
                  description: Se informado, verifica alertas reais dessa loja
      responses:
        '200':
          description: Teste enviado
        '400':
          description: Alertas não configurados
  /api/alertas/configurar:
    post:
      tags: [Alertas]
      summary: Atualizar configuração de alertas em runtime
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                chat_id: { type: string, example: "-1001234567890" }
                threshold: { type: number, example: 0.15 }
                apenas_quedas: { type: boolean }
      responses:
        '200':
          description: Configuração atualizada
  /api/coletar:
    post:
      tags: [Coleta]
      summary: Executar coleta e gravar snapshot
      description: |
        Dispara busca de produtos, aplica ranking, grava snapshot no BigQuery.
        Protegido por X-Garimpo-Token. Quando busca_id é informado e a fonte é shopee-shop,
        aplica amostragem rotativa (cursor de paginação) e throttling.
      security:
        - ColetaToken: []
      parameters:
        - name: fonte
          in: query
          schema: { type: string, enum: [shopee, shopee-shop, csv] }
        - name: keyword
          in: query
          schema: { type: string }
        - name: categoria
          in: query
          schema: { type: string }
        - name: estrategia
          in: query
          schema: { type: string, default: nicho }
        - name: top
          in: query
          schema: { type: integer, default: 20 }
        - name: shop_ids
          in: query
          schema: { type: string }
        - name: busca_id
          in: query
          description: Habilita rotação de catálogo e alertas automáticos
          schema: { type: string }
      responses:
        '202':
          description: Coleta executada
          content:
            application/json:
              schema:
                type: object
                properties:
                  categoria: { type: string }
                  estrategia: { type: string }
                  coletados: { type: integer }
                  em: { type: string, format: date-time }
        '401':
          description: Token inválido
  /api/publicar:
    post:
      tags: [Publicação]
      summary: Publicar oferta em canal (envio imediato)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PublicarRequest' }
      responses:
        '200':
          description: Publicação enviada
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Resultado' }
  /api/publicacoes:
    get:
      tags: [Publicação]
      summary: Lista publicações (agendadas, enviadas, erros)
      parameters:
        - name: status
          in: query
          schema: { type: string, enum: [agendada, enviada, erro, ""] }
      responses:
        '200':
          description: Lista de publicações
    post:
      tags: [Publicação]
      summary: Agendar ou enviar publicação
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PublicarRequest' }
      responses:
        '201':
          description: Publicação criada
  /api/publicar-pendentes:
    post:
      tags: [Publicação]
      summary: Enviar publicações agendadas cujo horário já passou
      security:
        - ColetaToken: []
      responses:
        '200':
          description: Resultado do processamento
          content:
            application/json:
              schema:
                type: object
                properties:
                  enviadas: { type: integer }
                  erros: { type: integer }
  /api/destinos:
    get:
      tags: [Destinos]
      summary: Lista destinos de publicação
      responses:
        '200':
          description: Lista de destinos
          content:
            application/json:
              schema:
                type: object
                properties:
                  destinos:
                    type: array
                    items: { $ref: '#/components/schemas/Destino' }
    post:
      tags: [Destinos]
      summary: Criar/atualizar destino
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Destino' }
      responses:
        '201':
          description: Destino salvo
    delete:
      tags: [Destinos]
      summary: Remover destino
      parameters:
        - name: id
          in: query
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Destino removido
  /api/templates:
    get:
      tags: [Templates]
      summary: Lista templates de mensagem
      responses:
        '200':
          description: Lista de templates
    post:
      tags: [Templates]
      summary: Criar/atualizar template
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Template' }
      responses:
        '201':
          description: Template salvo
    delete:
      tags: [Templates]
      summary: Remover template
      parameters:
        - name: id
          in: query
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Template removido
  /api/templates/preview:
    post:
      tags: [Templates]
      summary: Renderizar preview de template com dados
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                template_id: { type: string }
                nome: { type: string }
                preco: { type: number }
                categoria: { type: string }
      responses:
        '200':
          description: Preview renderizado
          content:
            application/json:
              schema:
                type: object
                properties:
                  preview: { type: string }
                  com_foto: { type: boolean }
  /api/estatisticas:
    get:
      tags: [Análise]
      summary: Estatísticas de mercado por categoria
      parameters:
        - name: dias
          in: query
          schema: { type: integer, default: 30 }
      responses:
        '200':
          description: Resumo por categoria
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Estatisticas' }
  /api/coletas:
    get:
      tags: [Análise]
      summary: Histórico de coletas executadas
      parameters:
        - name: dias
          in: query
          schema: { type: integer, default: 30 }
      responses:
        '200':
          description: Lista de coletas
  /api/conversoes:
    get:
      tags: [Análise]
      summary: Relatório de conversões por canal/sub_id
      parameters:
        - name: dias
          in: query
          schema: { type: integer, default: 30 }
      responses:
        '200':
          description: Conversões agrupadas
  /api/conversoes/sync:
    post:
      tags: [Coleta]
      summary: Sincroniza conversões reais da Shopee (conversionReport)
      description: |
        Consulta o conversionReport da API de afiliados da Shopee e retorna
        conversões dos últimos N dias. Chamado pelo Cloud Scheduler 1x/dia.
      security:
        - ColetaToken: []
      parameters:
        - name: dias
          in: query
          schema: { type: integer, default: 7 }
      responses:
        '200':
          description: Conversões sincronizadas
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string }
                  conversoes: { type: integer }
                  dias: { type: integer }
  /api/resolver-link:
    post:
      tags: [Curadoria]
      summary: Resolver link curto da Shopee para dados do produto
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                url: { type: string }
      responses:
        '200':
          description: Dados do produto extraídos
  /api/produto/origem:
    get:
      tags: [Curadoria]
      summary: Buscar origem e marca de um produto via API pública Shopee
      description: |
        Consulta a API pública v4 da Shopee (pdp/get_pc) para extrair os atributos
        "País de Origem" e "Marca" do produto. Resultado cacheado por 24h.
        Não requer autenticação. Usado pelo CandidateCard para enriquecer o badge.
      security: []
      parameters:
        - name: item_id
          in: query
          required: true
          schema: { type: string }
          example: "23720954312"
        - name: shop_id
          in: query
          required: true
          schema: { type: string }
          example: "785541033"
      responses:
        '200':
          description: Origem e marca do produto
          content:
            application/json:
              schema:
                type: object
                properties:
                  item_id: { type: string }
                  shop_id: { type: string }
                  origem: { type: string, description: "País normalizado (ex: Coreia, Japão, China)" }
                  marca: { type: string }
                  fonte: { type: string, enum: [api_publica, cache, erro] }
  /api/produto/origem/batch:
    post:
      tags: [Curadoria]
      summary: Buscar origem de múltiplos produtos (máx 20)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                itens:
                  type: array
                  items:
                    type: object
                    properties:
                      item_id: { type: string }
                      shop_id: { type: string }
      responses:
        '200':
          description: Resultados por produto
          content:
            application/json:
              schema:
                type: object
                properties:
                  resultados:
                    type: array
                    items:
                      type: object
                      properties:
                        item_id: { type: string }
                        shop_id: { type: string }
                        origem: { type: string }
                        marca: { type: string }
                        fonte: { type: string }
...
  /api/admin/logs:
    get:
      tags: [Admin]
      summary: Logs recentes do servidor
      parameters:
        - name: n
          in: query
          schema: { type: integer, default: 100 }
        - name: nivel
          in: query
          schema: { type: string, enum: [error, warn, info, debug, ""] }
      responses:
        '200':
          description: Logs e estatísticas
  /api/admin/log-level:
    post:
      tags: [Admin]
      summary: Alterar nível de log em runtime
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                nivel: { type: string, enum: [debug, info, warn, error] }
      responses:
        '200':
          description: Nível alterado
...
  /api/admin/shopee-introspect:
    get:
      tags: [Admin]
      summary: Introspecção do schema GraphQL da API Shopee
      description: |
        Faz 4 queries de introspecção na API de afiliados para descobrir campos
        disponíveis (especialmente relacionados a origem do produto). Restrito a admin.
      responses:
        '200':
          description: Resultados da introspecção
          content:
            application/json:
              schema:
                type: object
                properties:
                  app_id: { type: string }
                  endpoint: { type: string }
                  resultados:
                    type: array
                    items:
                      type: object
                      properties:
                        nome: { type: string }
                        sucesso: { type: boolean }
                        erro: { type: string }
                        resposta: { type: object }
                        campos_origem: { type: array, items: { type: string } }
                  campos_origem_global: { type: array, items: { type: string } }
                  conclusao: { type: string }
        '403':
          description: Não é admin
...
  /api/onboarding/termos:
    post:
      tags: [Onboarding]
      summary: "Step 1: Aceitar termos de uso (LGPD)"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [aceito]
              properties:
                aceito: { type: boolean, description: "Deve ser true" }
      responses:
        '200':
          description: Termos aceitos
          content:
            application/json:
              schema:
                type: object
                properties:
                  step: { type: integer }
                  status: { type: string }
  /api/onboarding/shopee:
    post:
      tags: [Onboarding]
      summary: "Step 2: Salvar credenciais Shopee Affiliate API"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [app_id, secret]
              properties:
                app_id: { type: string, example: "18332030606" }
                secret: { type: string, description: "Será criptografado antes de armazenar" }
      responses:
        '200':
          description: Credenciais salvas
        '403':
          description: Termos não aceitos
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Firebase ID Token
    ColetaToken:
      type: apiKey
      in: header
      name: X-Garimpo-Token
  schemas:
    Candidato:
      type: object
      properties:
        id: { type: string }
        nome: { type: string }
        categoria: { type: string }
        loja: { type: string, description: "Nome da loja vendedora" }
        origem: { type: string, description: "País de origem (ex: Coreia, Japão). Vazio se desconhecido." }
        preco: { type: number }
        comissao: { type: number, description: "Fração (0.15 = 15%)" }
        vendas: { type: integer }
        avaliacao: { type: number }
        score: { type: number, description: "Teor (0-1)" }
        link: { type: string, format: uri }
        imagem: { type: string, format: uri }
        suspeito: { type: boolean }
        exploracao: { type: boolean }
...
    Evento:
      type: object
      properties:
        tipo: { type: string, enum: [selecao, publicacao] }
        id: { type: string }
        nome: { type: string }
        categoria: { type: string }
        estrategia: { type: string }
        comissao: { type: number }
        preco: { type: number }
        vendas: { type: integer }
        score: { type: number }
...